backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.1
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0
redis>=5.0.1
scikit-learn>=1.5.0
pandas>=2.1.3
numpy>=1.26.2
joblib>=1.3.2
python-dotenv>=1.0.0
httpx>=0.25.2
psycopg2-binary>=2.9.9
prometheus-fastapi-instrumentator>=6.1.0
opensearch-py>=2.4.2
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.3.0
pydantic>=2.5.2
